+++ src/tests/utils.ts
@@ -1,0 +1,18 @@
+export async function isProcessRunning(processName: string): Promise<boolean> {
+    const cmd = (() => {
+      switch (process.platform) {
+        case 'win32': return `tasklist`
+        case 'darwin': return `ps -ax | grep ${processName}`
+        case 'linux': return `ps -A`
+        default: return false
+      }
+    })()
+  
+    return new Promise((resolve, reject) => {
+      require('child_process').exec(cmd, (err: Error, stdout: string, stderr: string) => {
+        if (err) reject(err)
+  
+        resolve(stdout.toLowerCase().indexOf(processName.toLowerCase()) > -1)
+      })
+    })
+  }
